pyproject.toml
[tool.poetry]
name = "studentdiary"
version = "0.1.0"
description = ""
authors = ["Alexandr Usenko <usenko.alexandr97@mail.ru>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest = "^7.2.2"
pytest-django = "^4.5.2"
django-extensions = "^3.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dj = "src.manage:main"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "src.studdiary.settings_test"
addopts = "--nomigrations -x"
testpaths = "src"
norecursedirs = "studdiary settings migrations templates views serializers admin mocks mock"
